requirements.txt
fastapi
uvicorn
python-dotenv
httpx
psycopg2
alembic